Mate slave wire

+++ AAEmu.Game/Core/Packets/C2G/CSSellHouseCancelPacket.cs
@@ -12,7 +12,7 @@ public class CSSellHouseCancelPacket : GamePacket
 
     public override void Read(PacketStream stream)
     {
-        var tl = stream.ReadUInt16();
+        var tl = (ushort)stream.ReadUInt32(); // the handle is 32 bits on the wire; ours is 16
 
         Logger.Debug("SellHouseCancel, Tl: {0}", tl);
         HousingManager.Instance.CancelForSale(tl, true);
