+++ AAEmu.Game/Core/Packets/G2C/SCHouseBuildPayChangedPacket.cs
@@ -5,10 +5,10 @@ namespace AAEmu.Game.Core.Packets.G2C;
 
 public class SCHouseBuildPayChangedPacket : GamePacket
 {
-    private readonly ushort _tl;
+    private readonly uint _tl;
     private readonly int _moneyAmount;
 
-    public SCHouseBuildPayChangedPacket(ushort tl, int moneyAmount) : base(SCOffsets.SCHouseBuildPayChangedPacket, 5)
+    public SCHouseBuildPayChangedPacket(uint tl, int moneyAmount) : base(SCOffsets.SCHouseBuildPayChangedPacket, 5)
     {
         _tl = tl;
         _moneyAmount = moneyAmount;
