+++ AAEmu.Game/Core/Packets/G2C/SCHousingRecoverTogglePacket.cs
@@ -5,10 +5,10 @@ namespace AAEmu.Game.Core.Packets.G2C;
 
 public class SCHousingRecoverTogglePacket : GamePacket
 {
-    private readonly ushort _tl;
+    private readonly uint _tl;
     private readonly bool _allowRecover;
 
-    public SCHousingRecoverTogglePacket(ushort tl, bool allowRecover) : base(SCOffsets.SCHousingRecoverTogglePacket, 5)
+    public SCHousingRecoverTogglePacket(uint tl, bool allowRecover) : base(SCOffsets.SCHousingRecoverTogglePacket, 5)
     {
         _tl = tl;
         _allowRecover = allowRecover;
